描画周りを細かく修正
